Fix some type conversion warnings

+++ src/network/network_server.cpp
@@ -474,7 +474,7 @@ NetworkRecvStatus ServerNetworkGameSocketHandler::SendDesyncLog(const std::strin
 	for (size_t offset = 0; offset < log.size();) {
 		Packet *p = new Packet(PACKET_SERVER_DESYNC_LOG, SHRT_MAX);
 		size_t size = std::min<size_t>(log.size() - offset, SHRT_MAX - 2 - p->Size());
-		p->Send_uint16(size);
+		p->Send_uint16(static_cast<uint16>(size));
 		p->Send_binary(log.data() + offset, size);
 		this->SendPacket(p);
